+++ package.json
@@ -11,10 +11,9 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.2.0",
     "@testing-library/user-event": "^14.1.1",
-    "@types/jest": "^27.5.0",
     "@types/node": "^17.0.32",
     "@types/react": "^18.0.9",
-    "@types/react-dom": "^18.0.3",
+    "@types/react-dom": "^18.0.4",
     "@types/redux-logger": "^3.0.9",
     "axios": "^0.27.2",
     "node-fetch": "^3.2.4",
@@ -56,5 +55,9 @@
     ]
   },
   "type": "module",
-  "exports": "./index.js"
+  "exports": "./index.js",
+  "devDependencies": {
+    "@types/jest": "^27.5.1",
+    "@types/mocha": "^9.1.1"
+  }
 }
