--- src/App.test.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-import { Provider } from 'react-redux';
-import App from './App';
-
-test('renders learn react link', () => {
-  const { getByText } = render(
-      <App />
-  );
-
-  expect(getByText(/learn/i)).toBeInTheDocument();
-});
