+++ src/App.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Dashboard from './component/pages/Dashboard';
-import Customer from './component/pages/User'
+import Customer from './component/pages/Customer'
 import AddUser from './component/pages/AddUser';
 import EditUser from './component/pages/EditUser';
+import SignIn from './component/Formlogin';
+import SignUp from './component/Formsignup';
 
 const App = () => {
 
@@ -11,7 +12,9 @@ const App = () => {
     <div className='App'>
       <BrowserRouter>
         <Routes>
-          <Route path='/' element={<Customer/>}/>
+          <Route path="/" element={<SignIn/>}/>
+          <Route path="/SignUp" element={<SignUp/>}/>
+          <Route path='/Customer' element={<Customer/>}/>
           <Route path='/AddUser' element={<AddUser/>}/>
           <Route path='/EditUser/:id' element={<EditUser/>}/>
         </Routes>
