+++ src/component/pages/user.css
@@ -5,11 +5,20 @@
 table {
   padding: 0px;
   margin-top: 100px;
-  width: 70%;
+  width: 80%;
 }
-/* button {
+#drawer {
+  width: 210px;
+}
+.list {
   position: relative;
-  width: 60px;
-  height: 30px;
-  left: 80px;
-} */
+  top: -350px;
+}
+.add-user {
+  position: relative;
+  top: 60px;
+  left: 17px;
+}
+.table {
+  margin-left: 10px;
+}
