+++ src/reduxthunk/userAction.ts
@@ -2,8 +2,7 @@ import { Dispatch } from 'redux';
 import { AppActions } from './userActiontype';
 import {  FETCH_USER_REQUEST,
           FETCH_USER_SUCCESS,
-          FETCH_USER_FAILURE,
-          DELETE_USER_ACTION } from './userActiontype';
+        } from './userActiontype';
 import { User } from './userActiontype';
 
 const url = "http://localhost:3001/users"
@@ -59,17 +58,21 @@ export const deleteUserAction = (id: number) => {
 
 export const getUserId = (id: number) => {
   return (dispatch: Dispatch<AppActions>) => {
+      dispatch(requestUser())
       fetch(url+'/'+id)
       .then((res) => res.json())
-      .then((json) => dispatch(receiveUser(json)));
+      .then((json) => {
+        console.log("jsonnn", json)
+        dispatch(receiveUser(json))
+      });
   }
 
 }
 
-export const editUserAction = (id: string, user: User) => {
+export const editUserAction = (id: number, user: User) => {
   return (dispatch: Dispatch<AppActions>) => {
     var options = {
-      method: "PATCH",
+      method: "PUT",
       body: JSON.stringify(user),
       headers: {
         'Content-type': 'application/json; charset=UTF-8'
