+++ tsconfig.json
@@ -18,7 +18,11 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "noEmit": true,
-    "jsx": "preserve"
+    "jsx": "preserve",
+    "types": [
+      "react/next",
+      "react-dom/next"
+    ]
   },
   "include": [
     "src/**/*"
